+++ intercom_flutter_web/CHANGELOG.md
@@ -1,20 +1,29 @@
 # Changelog
 
+# 0.1.0
+
+- Added method `loginIdentifiedUser` with `IntercomStatusCallback` support.
+- Deprecated `registerIdentifiedUser` in favor of `loginIdentifiedUser`.
+- Added method `loginUnidentifiedUser` with `IntercomStatusCallback` support.
+- Deprecated `registerUnidentifiedUser` in favor of `loginUnidentifiedUser`.
+- Added parameter `statusCallback` in updateUser to support `IntercomStatusCallback`.
+- Updated `intercom_flutter_platform_interface` version to `1.1.0`.
+
 # 0.0.5
 
-- implemented `displayArticle` [(showArticle)](https://developers.intercom.com/installing-intercom/docs/intercom-javascript#intercomshowarticle-articleid)
+- Implemented `displayArticle` [(showArticle)](https://developers.intercom.com/installing-intercom/docs/intercom-javascript#intercomshowarticle-articleid)
 
 # 0.0.4
 
-- updated dependency `intercom_flutter_platform_interface: ^1.0.1`
+- Updated dependency `intercom_flutter_platform_interface: ^1.0.1`
 
 # 0.0.3
 
-- resolved issue [#173](https://github.com/v3rm0n/intercom_flutter/issues/173)
+- Resolved issue [#173](https://github.com/v3rm0n/intercom_flutter/issues/173)
 
 # 0.0.2
 
-- updated dependency intercom_flutter_platform_interface: ^1.0.0
+- Updated dependency intercom_flutter_platform_interface: ^1.0.0
 
 # 0.0.1